#59 refactor: Exception 요구사항 변경
